add safety checks

+++ src/wp-admin/includes/file.php
@@ -1934,12 +1934,18 @@ function copy_dir( $from, $to, $skip_list = array() ) {
 function move_dir( $from, $to ) {
 	global $wp_filesystem;
 
-	$wp_filesystem->rmdir( $to );
-	if ( @rename( $from, $to ) ) {
-		return true;
+	if ( 'direct' === $wp_filesystem->method ) {
+		$wp_filesystem->rmdir( $to );
+		if ( @rename( $from, $to ) ) {
+			return true;
+		}
 	}
 
-	$wp_filesystem->mkdir( $to );
+	if ( ! $wp_filesystem->is_dir( $to ) ) {
+		if ( ! $wp_filesystem->mkdir( $to, FS_CHMOD_DIR ) ) {
+			return new WP_Error( 'mkdir_failed_copy_dir', __( 'Could not create directory.' ), $to );
+		}
+	}
 	$result = copy_dir( $from, $to );
 
 	return $result;
